requirements.txt
PyQt6>=6.4.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
numpy>=1.24.0
Pillow>=10.0.0
